Move logic to get docs.json to ProjectJsonFiles
