+++ lib/project-dependencies.js
@@ -1,13 +1,6 @@
 const path = require('path');
-const util = require('util');
 const os = require('os');
-const got = require('got');
-const fs = require('fs-extra');
 const ProjectJsonFiles = require('./project-json-files');
-
-const fsMkdirp = util.promisify(fs.mkdirp);
-const fsReadJson = util.promisify(fs.readJson);
-const fsWriteJson = util.promisify(fs.writeJson);
 
 const elmRoot =
   process.env.ELM_HOME ||
@@ -52,7 +45,7 @@ async function collect(options, elmJson, reviewDirectDependencies, elmVersion) {
       const packageVersion = constraint.split(' ')[0];
 
       const [docsJson, dependencyElmJson] = await Promise.all([
-        getDocsJson(
+        ProjectJsonFiles.getDocsJson(
           directory(name, packageVersion),
           elmReviewDependencyCache(name, packageVersion),
           name,
@@ -155,31 +148,6 @@ function linkToModule(dependencyName, packageVersion, moduleName) {
   return `https://package.elm-lang.org/packages/${dependencyName}/${packageVersion}/${urlModuleName}`;
 }
 
-function getDocsJson(
-  directory,
-  elmReviewDependencyCache,
-  name,
-  packageVersion
-) {
-  return fsReadJson(path.join(directory, 'docs.json'))
-    .catch(() => fsReadJson(path.join(elmReviewDependencyCache, 'docs.json')))
-    .catch(() =>
-      readDocsJsonFromPackagesWebsite(name, packageVersion).then((result) => {
-        cacheFile(
-          elmReviewDependencyCache,
-          'docs.json',
-          result
-        ).catch(() => {});
-        return result;
-      })
-    );
-}
-
-async function cacheFile(directory, name, jsonContent) {
-  await fsMkdirp(directory).catch(() => {});
-  return fsWriteJson(path.join(directory, name), jsonContent);
-}
-
 function defaultElmJson(name, packageVersion) {
   return {
     type: 'package',
@@ -195,13 +163,6 @@ function defaultElmJson(name, packageVersion) {
     },
     'test-dependencies': {}
   };
-}
-
-async function readDocsJsonFromPackagesWebsite(packageName, packageVersion) {
-  const response = await got(
-    `https://package.elm-lang.org/packages/${packageName}/${packageVersion}/docs.json`
-  );
-  return JSON.parse(response.body);
 }
 
 module.exports = {
