+++ lib/project-json-files.js
@@ -63,11 +63,6 @@ function getElmJsonFromElmHome(elmVersion, name, packageVersion) {
   return promise;
 }
 
-async function cacheFile(directory, name, jsonContent) {
-  await fsMkdirp(directory).catch(() => {});
-  return fsWriteJson(path.join(directory, name), jsonContent);
-}
-
 async function readElmJsonFromPackagesWebsite(packageName, packageVersion) {
   const response = await got(
     `https://package.elm-lang.org/packages/${packageName}/${packageVersion}/elm.json`
@@ -75,7 +70,40 @@ async function readElmJsonFromPackagesWebsite(packageName, packageVersion) {
   return JSON.parse(response.body);
 }
 
+function getDocsJson(
+  directory,
+  elmReviewDependencyCache,
+  name,
+  packageVersion
+) {
+  return fsReadJson(path.join(directory, 'docs.json'))
+    .catch(() => fsReadJson(path.join(elmReviewDependencyCache, 'docs.json')))
+    .catch(() =>
+      readDocsJsonFromPackagesWebsite(name, packageVersion).then((result) => {
+        cacheFile(
+          elmReviewDependencyCache,
+          'docs.json',
+          result
+        ).catch(() => {});
+        return result;
+      })
+    );
+}
+
+async function readDocsJsonFromPackagesWebsite(packageName, packageVersion) {
+  const response = await got(
+    `https://package.elm-lang.org/packages/${packageName}/${packageVersion}/docs.json`
+  );
+  return JSON.parse(response.body);
+}
+
+async function cacheFile(directory, name, jsonContent) {
+  await fsMkdirp(directory).catch(() => {});
+  return fsWriteJson(path.join(directory, name), jsonContent);
+}
+
 module.exports = {
   getElmJson,
-  getElmJsonFromElmHome
+  getElmJsonFromElmHome,
+  getDocsJson
 };
